bash script now should work

+++ install.sh
@@ -20,11 +20,13 @@ cd "$1"
 CHECK_MK_LOCAL_PATH="/usr/local/lib/check_mk_agent"
 
 mkdir -p $CHECK_MK_LOCAL_PATH/local
-cp agents/check_mk_agent.macos $CHECK_MK_LOCAL_PATH
+cp agents/check_mk_agent.macosx $CHECK_MK_LOCAL_PATH
 cp -r agents/plugins/* $CHECK_MK_LOCAL_PATH
 cp LaunchDaemon/de.mathias-kettner.check_mk.plist /Library/LaunchDaemons/
-ln -s $CHECK_MK_LOCAL_PATH/check_mk_agent.macosx /usr/local/bin/check_mk_agent
-mkdir /etc/check_mk
+if [ ! -e /usr/local/bin/check_mk_agent ]; then
+	ln -s $CHECK_MK_LOCAL_PATH/check_mk_agent.macosx /usr/local/bin/check_mk_agent
+fi
+mkdir -p /etc/check_mk
 
 touch /var/run/de.arts-others.softwareupdatecheck
 touch /var/log/check_mk.err
@@ -34,5 +36,3 @@ chmod o+rw /var/run/de.arts-others.softwareupdatecheck
 chmod 666 /var/log/check_mk.err
 chown -R root:admin $CHECK_MK_LOCAL_PATH
 chmod 644 /Library/LaunchDaemons/de.mathias-kettner.check_mk.plist
-
-launchctl load -w /Library/LaunchDaemons/de.mathias-kettner.check_mk.plist
